refactor: 컨트롤러 메서드 외부로 분리
